chore: move vcs service handler from querier to query frontend
