+++ pkg/frontend/vcs/encryption.go
