+++ pkg/frontend/vcs/encryption_test.go
